Switch from timetk to slider for sliding window
`timetk` got archived from CRAN at one point, causing this package to get archived as well.

While `timetk` is back up, we only use it for a single function, and that function is basically a re-export from the lower-level `slider` package.

So this removes our dependency on `timetk` and replaces it with a dependency on `slider` instead.

As it turns out, this also gives us better results, since we're able to more clearly note that we're looking for a sliding window on the "time" column of 30 seconds. So the test is updated, and the results look more correct.

Resolves #2